reorganize the package
