+++ woodelf/__init__.py
@@ -1,7 +1,4 @@
-import woodelf.cube_metric
-import woodelf.decision_trees_ensemble
-import woodelf.parse_models
-import woodelf.path_to_matrices
+import woodelf.core.cube_metric
 import woodelf.simple_woodelf
 from woodelf.explainer import WoodelfExplainer
 from woodelf.pdp_display import WoodelfPartialDependenceDisplay
