+++ woodelf/core/lts_polynomial_multiplication.py
