+++ woodelf/core/trees/parse_models.py
@@ -1,5 +1,5 @@
-from woodelf.decision_trees_ensemble import DecisionTreeNode, LeftIsSmallerEqualDecisionTreeNode, DecisionTreesEnsemble
-from woodelf.utils import safe_isinstance
+from woodelf.core.trees.decision_trees_ensemble import DecisionTreeNode, LeftIsSmallerEqualDecisionTreeNode, DecisionTreesEnsemble
+from woodelf.core.trees.parsing_utils import safe_isinstance
 
 MODEL_CLASS_TO_DECISION_TREE_CLASS = {
     # sklearn regressors
